src/requirements.txt
twilio
Pillow
python-dotenv
